Moved files to data folder for cleaner directory structure
