+++ src/main/java/com/myuuiii/empirewandplus/SpellEffects/Fireball/FireballEffect.java
@@ -3,8 +3,8 @@ package com.myuuiii.empirewandplus.SpellEffects.Fireball;
 import com.myuuiii.empirewandplus.Abstracts.Spell;
 import com.myuuiii.empirewandplus.EmpireWandPlus;
 import com.myuuiii.empirewandplus.SpellEffects.SpellEffect;
-import com.myuuiii.empirewandplus.SpellEntityLists;
-import com.myuuiii.empirewandplus.SpellNames;
+import com.myuuiii.empirewandplus.Data.SpellEntityLists;
+import com.myuuiii.empirewandplus.Data.SpellNames;
 import org.bukkit.Particle;
 import org.bukkit.Sound;
 import org.bukkit.entity.Entity;
