+++ src/main/java/com/myuuiii/empirewandplus/SpellEffects/Pulse/EmpirePulseEffect.java
@@ -3,8 +3,8 @@ package com.myuuiii.empirewandplus.SpellEffects.Pulse;
 import com.myuuiii.empirewandplus.Abstracts.Spell;
 import com.myuuiii.empirewandplus.EmpireWandPlus;
 import com.myuuiii.empirewandplus.SpellEffects.SpellEffect;
-import com.myuuiii.empirewandplus.SpellEntityLists;
-import com.myuuiii.empirewandplus.SpellNames;
+import com.myuuiii.empirewandplus.Data.SpellEntityLists;
+import com.myuuiii.empirewandplus.Data.SpellNames;
 import org.bukkit.Color;
 import org.bukkit.FireworkEffect;
 import org.bukkit.Particle;
