+++ src/main/java/com/myuuiii/empirewandplus/Spells/Fireball/Fireball.java
@@ -1,8 +1,8 @@
 package com.myuuiii.empirewandplus.Spells.Fireball;
 
 import com.myuuiii.empirewandplus.Abstracts.RangedSpell;
-import com.myuuiii.empirewandplus.SpellEntityLists;
-import com.myuuiii.empirewandplus.SpellValues;
+import com.myuuiii.empirewandplus.Data.SpellEntityLists;
+import com.myuuiii.empirewandplus.Data.SpellValues;
 import org.bukkit.Location;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.Player;
